(qa) Fix reboot and shutdown quick actions on newer systems

+++ pulse_xmpp_agent/lib/utils.py
@@ -115,6 +115,7 @@ if sys.platform.startswith("linux"):
     import grp
     import posix_ipc
     import xmltodict
+    import distro
 
 if sys.platform.startswith("darwin"):
     import pwd
@@ -2455,40 +2456,26 @@ if sys.platform.startswith("win"):
             return None
 
 
-def shutdown_command(time=0, msg=""):
+def shutdown_command():
     """
-    This  function allow to shutdown a machine, and if needed
-    to display a message
-
-    Args:
-        time: the delay before the shutdown
-        msg:  the message that will be displayed
-
+    This function allows to shutdown a machine
     """
-    if msg != "":
-        msg = msg.strip('" ')
-        msg = f'"{msg}"'
     if sys.platform.startswith("linux"):
-        if int(time) == 0 or msg == "":
-            cmd = "shutdown now"
+        if distro.id().lower() in [
+            "ubuntu",
+            "zorin",
+            "linuxmint",
+            "debian",
+        ]:
+            cmd = 'systemctl poweroff -i'
         else:
-            cmd = f"shutdown -P -f -t {time} {msg}"
-        logging.debug(cmd)
-        os.system(cmd)
+            cmd = 'shutdown -P now'
     elif sys.platform.startswith("win"):
-        if int(time) == 0 or msg == "":
-            cmd = "shutdown /p"
-        else:
-            cmd = f"shutdown /s /t {time} /c {msg}"
-        logging.debug(cmd)
-        os.system(cmd)
+        cmd = 'shutdown /p'
     elif sys.platform.startswith("darwin"):
-        if int(time) == 0 or msg == "":
-            cmd = "shutdown -h now"
-        else:
-            cmd = f'shutdown -h +{time} "{msg}"'
-        logging.debug(cmd)
-        os.system(cmd)
+        cmd = 'shutdown -h now'
+    logging.debug(cmd)
+    os.system(cmd)
     return
 
 
@@ -2517,7 +2504,15 @@ def reboot_command():
     This function allow to reboot a machine.
     """
     if sys.platform.startswith("linux"):
-        os.system("shutdown -r now")
+        if distro.id().lower() in [
+            "ubuntu",
+            "zorin",
+            "linuxmint",
+            "debian",
+        ]:
+            os.system("systemctl reboot -i")
+        else:
+            os.system("shutdown -r now")
     elif sys.platform.startswith("win"):
         os.system("shutdown /r")
     elif sys.platform.startswith("darwin"):
